+++ app/fdroid-repository-android/src/main/java/de/k3b/fdroid/android/repository/VersionDao.java
@@ -45,7 +45,7 @@ public interface VersionDao extends VersionRepository {
 
     @Query("SELECT App_Version.* FROM App_Version inner join App on App_Version.appId = App.id " +
             "WHERE App.repoId = :repoId AND App.packageName = :packageName AND App_Version.versionCode = :versionCode")
-    Version findByRepoPackageAndVersionCode(int repoId, String packageName, long versionCode);
+    Version findByRepoPackageAndVersionCode(int repoId, String packageName, int versionCode);
 
     @Query(value = "select al.* from App_Version al " +
             "where al.appId in (:appIds) " +
